+++ NWN.Anvil/src/main/Internal/Assemblies.cs
@@ -7,7 +7,7 @@ namespace Anvil.Internal
 {
   internal static class Assemblies
   {
-    internal const string TargetFramework = "net5.0";
+    internal static readonly string[] TargetFrameworks = { "net6.0", "net5.0" };
 
     internal static readonly Assembly Anvil = typeof(Assemblies).Assembly;
     internal static readonly Assembly Core = typeof(NWN.Core.NWNCore).Assembly;